refactor: move user-request-queue to it's own file
